minor update
Added a playable area scene to make adding a playable zone easier in 2d without depending on tilemaps. It is a rectangle that has a background, toggle borders, and an area2d to detects if physics bodies enter or exit the area. It is a static body for now, though that makes the naming misleading despite it acts like an area as well as a static body. I may change since it may make use of the area signals more than the static body or I could make it a node2d and add more properties (I rather not).