minor update
I been working on how actions will work. I have dedicated action (like sprint) and abstract action (like basic and secondary attacks as well as abilities). Also I am catching when the input us pressed (controller is holding it, but character is setting it though the handler may also set it depending on which is more reliable. It is used for checking what action is still active. I still need to add a way to check and block or cancel action base on this state. Most likely the character responsibility).